Refactor : /user/address -> /auth/address 변경 및 조회 코드 리팩터링
