+++ src/main/java/com/mentionall/cpr2u/user/controller/UserController.java
@@ -28,18 +28,6 @@ public class UserController {
 
     private final AddressService addressService;
 
-    @Operation(summary = "주소 리스트 조회", description = "전국 시도와 시군구 주소 리스트를 조회한다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공", content = @Content(array = @ArraySchema(schema = @Schema(implementation = AddressResponseDto.class)))),
-    })
-    @GetMapping("/address")
-    public ResponseEntity<ResponseDataTemplate> readAddressList() {
-        return ResponseDataTemplate.toResponseEntity(
-                OK_SUCCESS,
-                addressService.readAll()
-        );
-    }
-
     @Operation(summary = "사용자 주소 설정", description = "사용자의 주소를 파라미터로 넘어온 ID의 주소로 설정한다.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "성공", content = @Content(array = @ArraySchema(schema = @Schema(implementation = ResponseTemplate.class)))),
